+++ lib/prawn/swiss_qr_bill/iban.rb
@@ -2,6 +2,7 @@
 
 module Prawn
   module SwissQRBill
+    class MissingIBANError < StandardError; end
     class InvalidIBANError < StandardError; end
 
     # Check validity of IBAN.