+++ lib/prawn/swiss_qr_bill/sections/qr_code.rb
@@ -71,6 +71,8 @@ module Prawn
         end
 
         def generate_qr_data(data)
+          validate(data) if @options[:validate]
+
           flat_data = {}
           MAPPING.each_key do |key|
             # check if the exists
@@ -80,10 +82,25 @@ module Prawn
           end
 
           iban = IBAN.new(data[:creditor][:iban])
-          raise InvalidIBANError, "IBAN #{iban.prettify} is invalid" if @options[:validation] && !iban.valid?
 
           qr_data = QR::Data.new(flat_data.merge(iban: iban.code))
           qr_data.generate
+        end
+
+        def validate(data)
+          # IBAN must be given
+          raise MissingIBANError, 'IBAN is missing' unless data[:creditor][:iban]
+
+          iban = IBAN.new(data[:creditor][:iban])
+
+          raise InvalidIBANError, "IBAN #{iban.prettify} is invalid" unless iban.valid?
+
+          # reference may be missing (optional)
+          return unless data[:reference]
+
+          ref = Reference.new(data[:reference])
+
+          raise InvalidReferenceError, "Reference #{@data[:reference]} is invalid" unless ref.valid?
         end
 
         # hash: hash to test key path for